pyproject.toml
[project]
name = "omi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all,standard]>=0.120.4",
    "jsonrpcclient>=4.0.3",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
]
